user-service/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
pika==1.3.1
PyJWT==2.3.0
pymongo==4.3.3
requests==2.25.1
